requirements/base.txt
django==3.0.6
attrs==19.3.0
djangorestframework==3.11.0
